without chrome driver and remove audio file issue

+++ AP_DR/src/main/java/com/apdr/utalities/WaitUtils.java
@@ -7,5 +7,5 @@ public class WaitUtils extends TestBase {
   
 	public static long PAGE_LOAD_TIMEOUT=50;
     public static long IMPLICIT_WAIT = 20;
-    public static long Duration = 60;
+    public static int Duration = 60;
 }
